worker/requirements.txt
Flask==2.2.2
Flask-RESTful==0.3.9
Flask-SQLAlchemy==3.0.3
flask-marshmallow==0.14.0
marshmallow==3.19.0
marshmallow-enum==1.5.1
marshmallow-sqlalchemy==0.28.1
SQLAlchemy==2.0.2
psycopg2-binary==2.9.5
Werkzeug==2.2.2
gunicorn
numpy==1.26.4
opencv-python==4.9.0.80
requests==2.28.1
google-cloud-storage==2.16.0
google-cloud-pubsub